api (fix): only try to delete tmp image if keep images is enabled

+++ api/deletePhoto.php
@@ -41,11 +41,13 @@ if ($config['chroma_keying']) {
     }
 }
 
-if (is_readable($filePathTmp)) {
-    if (!unlink ($filePathTmp)) {
-        die(json_encode([
-            'error' => 'Could not delete tmp file',
-        ]));
+if ($config['keep_images']) {
+    if (is_readable($filePathTmp)) {
+        if (!unlink ($filePathTmp)) {
+            die(json_encode([
+                'error' => 'Could not delete tmp file',
+            ]));
+        }
     }
 }
